Adaugare problema ultimacifraapatiintregic

+++ a16.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-    long long a;
+    long a;
     cin >> a;
     a = a * a;
     a = a * a;
